Enable to create more bucketizations over a pcap, rework datasource ui for this, rework query execution ui model, add dashboard state

+++ backend/src/main/kotlin/com/netdash/rest/pcap/resource/PcapParserResource.kt
@@ -1,5 +1,7 @@
 package com.netdash.rest.pcap.resource
 
+import com.netdash.rest.pcap.bucket.model.BucketizationRequest
+import com.netdash.rest.pcap.bucket.service.BucketService
 import com.netdash.rest.pcap.model.PcapDescriptor
 import com.netdash.rest.pcap.repository.PcapRepositoryFactory
 import com.netdash.rest.pcap.repository.TableIdentifier
@@ -16,18 +18,29 @@ class PcapParserResource(
     private val fileService: FileService,
     private val pcapRepositoryFactory: PcapRepositoryFactory,
     private val allPcapRepository: JdbcPcapRepository,
+    private val bucketService: BucketService
 ) {
 
     @PostMapping
     fun parsePcapFile(
         @RequestParam("file") file: MultipartFile,
-        @RequestParam("type") tableIdentifier: String,
+        @RequestParam("type") tableIdentifier: String
     ): ResponseEntity<Int> {
+        if (file.originalFilename == null)
+            return ResponseEntity.badRequest().build();
         val path = fileService.store(file)
         val pcapData = pcapParserService.parse(path, file.originalFilename!!)
         pcapRepositoryFactory.get(TableIdentifier.from(tableIdentifier)).save(pcapData)
         fileService.delete(path)
         return ResponseEntity.ok(pcapData.pcapData.size)
+    }
+
+    @PostMapping("/bucket")
+    fun bucketize(
+        @RequestBody bucketizationRequest: BucketizationRequest
+    ): ResponseEntity<Long> {
+        val bucketCount = bucketService.bucketize(bucketizationRequest)
+        return ResponseEntity.ok(bucketCount)
     }
 
     @GetMapping
